ref(replay): Refactor ReplayReader to provide eventsWithSnapshots and touchEvents lists for video replays
